rename: component 폴더명 수정
